Fix (<BusinessDetails />): Move <BusinessDetails /> and remove external CSS styling
