--- client/src/pages/BusinessDetails/businessDetails.css
@@ -1,55 +1,0 @@
-.bus-details {
-  display: flex;
-  flex-direction: column;
-  place-items: center;
-}
-
-.bus-details--container {
-  display: flex;
-  width: 90vw;
-}
-
-.bus-details--info {
-  margin-left: 0.5rem;
-}
-
-/*! ----- Media Queries Begin ---- */
-/* Tablet View */
-@media screen and (min-width: 768px) {
-  .bus-details--header {
-    display: flex;
-    place-items: stretch;
-  }
-  .content-header {
-    width: 100vw;
-    padding: 0;
-  }
-  .bus-details--container {
-    width: 98%;
-    justify-content: space-between;
-  }
-  .bus-details--button {
-    width: 6rem;
-    height: 2rem;
-  }
-}
-
-/* Desktop Views */
-@media screen and (min-width: 1024px) {
-  .content-header {
-    width: 76vw;
-  }
-}
-
-@media screen and (min-width: 1400px) {
-  .content-header {
-    width: 82vw;
-  }
-}
-
-/* 4K View */
-@media screen and (min-width: 2560px) {
-  .content-header {
-    width: 89vw;
-  }
-}
